tests/wpt/web-platform-tests/css/css-typed-om/stylevalue-subclasses/cssUrlImageValue-invalid.html

<!doctype html>
<meta charset="utf-8">
<title>CSSURLImageValue Error Handling</title>
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#imagevalue-objects">
<meta name="assert" content="Test CSSURLImageValue constructor error handling" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<div id="log">
<script>
'use strict';

test(() => {
  assert_throws(new TypeError(), () => new CSSURLImageValue("file://:invalid url"));
}, 'Constructing a CSSURLImageValue with an invalid URL throws a TypeError');

</script>
